requirements.txt
dill == 0.3.6
numpy == 1.22.3
pandas == 2.0.2 
torch-geometric == 2.3.1
cdt == 0.6.0
dowhy ==  0.10.1
statsmodels == 0.14.0